feat: migrate authentication to Clerk and remove Supabase integration from login and signup pages
