+++ src/app/login/[[...rest]]/page.tsx
@@ -1,0 +1,14 @@
+import { SignIn } from '@clerk/nextjs';
+
+export default function LoginPage() {
+  return (
+    <div className="flex min-h-screen items-center justify-center bg-background px-4">
+      {/* Background pattern */}
+      <div className="fixed inset-0 bg-[linear-gradient(to_right,hsl(var(--border)/0.2)_1px,transparent_1px),linear-gradient(to_bottom,hsl(var(--border)/0.2)_1px,transparent_1px)] bg-[size:4rem_4rem]" />
+      <div className="fixed inset-0 bg-[radial-gradient(ellipse_at_center,transparent_20%,hsl(var(--background))_70%)]" />
+      <div className="relative z-10">
+        <SignIn />
+      </div>
+    </div>
+  );
+}
